+++ src/theme/components/header/headerLogin.ts
@@ -76,6 +76,15 @@ const HeaderLogin = {
         ...loginFocus(),
         borderRadius: { base: "0", md: "28px" },
       },
+      _dark: {
+        color: "ui.gray.xx-dark",
+        svg: {
+          fill: "ui.gray.xx-dark",
+        },
+        _hover: {
+          color: "ui.gray.xx-dark",
+        },
+      },
     },
     patronGreeting: {
       alignSelf: "flex-start",
@@ -120,6 +129,15 @@ const HeaderLogin = {
       _focus: {
         ...loginFocus(),
         borderRadius: { base: "0", md: "28px !important" },
+      },
+      _dark: {
+        color: "ui.gray.xx-dark",
+        svg: {
+          fill: "ui.gray.xx-dark",
+        },
+        _hover: {
+          color: "ui.gray.xx-dark",
+        },
       },
     },
   }),